Move semaphore functions and wrapper class to `base/sphore.cpp/h`
Move `sphore_*` functions from `base/system.cpp/h` and `CSemaphore` class from `base/tl/threading.h` to `base/sphore.cpp/h`.
