+++ src/engine/client/backend_sdl.cpp
@@ -6,8 +6,8 @@
 
 #include <base/log.h>
 #include <base/math.h>
+#include <base/sphore.h>
 #include <base/thread.h>
-#include <base/tl/threading.h>
 
 #include <engine/shared/config.h>
 #include <engine/shared/localization.h>
