+++ src/engine/shared/jobs.cpp
@@ -2,6 +2,8 @@
 /* If you are missing that file, acquire a complete release at teeworlds.com.                */
 #include "jobs.h"
 
+#include <base/dbg.h>
+#include <base/system.h> // TODO: replace with str.h after moving str_format
 #include <base/thread.h>
 
 #include <algorithm>
