+++ src/test/score_test.cpp
@@ -1,4 +1,6 @@
 #include <base/detect.h>
+#include <base/system.h> // TODO: replace with str.h after moving str_format
+#include <base/time.h>
 
 #include <engine/server/databases/connection.h>
 #include <engine/server/databases/connection_pool.h>
