+++ src/test/thread_test.cpp
@@ -1,7 +1,7 @@
 #include <base/lock.h>
+#include <base/sphore.h>
 #include <base/system.h>
 #include <base/thread.h>
-#include <base/tl/threading.h>
 
 #include <gtest/gtest.h>
